Refactor HTTP client and header management
Modularize HTTP client functionality with a new package `httpwrap` and improve header handling. This update includes the creation of dedicated header and client structures, as well as the move of authorization and error handling. Relevant methods and functionalities are migrated to ensure streamlined HTTP requests and improved code maintainability.

created auth package that now has most of the code and is now called by the scraper.

There is still a ways to go here but I'm trying to do this without changing the libraries API so we can test it and then refactor that part.

I changed my mind on strategy several times because it got so broke I couldn't make sense of anything.